feat: modelo de datos completo - 16 tablas creadas en MySQL
Todas las entidades del marketplace listas

Convención global para nombrar FKs = 'IdTabla'

Protección contra borrado en cascada accidental (Restrict en todas las FKs)

Primera migración aplicada y funcionando

+++ src/MarketHub.Domain/Entities/Usuario.cs
@@ -5,7 +5,7 @@ namespace MarketHub.Domain.Entities
     public class Usuario
     {
         public int Id { get; set; }
-        public long IdPerfil { get; set; }
+        public int IdPerfil { get; set; }
         public string Nombre { get; set; } = string.Empty;
         public string Apellido { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
